+++ changelog.md
@@ -5,6 +5,7 @@
   * Changed log level from INFO to DEBUG when sys.menu is not found
   * Enhancement: saveTests creates a new db record and uses the testUpdated NCube metaproperty.
   * Updated NCube.duplicate() to include ApplicationID
+  * Update hsqldb (2.4.0) and mysql (6.0.6) dependencies
 * 4.0.5
   * Support mode for json format when getting raw json (json-pretty, json)
   * Update to Spring Boot 1.5.3, Spring 4.3.8, and Groovy 2.4.11